make group data default, remove individual data
