+++ client/src/app/recurrence/form.jsx
@@ -17,13 +17,13 @@ import { useStore, defaultParams } from "./store";
 
 const FileInput = dynamic(() => import("@/components/file-input"), { ssr: false });
 
-export default function IndividualDataForm({ id }) {
+export default function GroupDataForm({ id }) {
   const queryClient = useQueryClient();
   const router = useRouter();
   const pathname = usePathname();
   const setState = useStore((state) => state.setState);
   const resetStore = useStore((state) => state.resetStore);
-  const individualData = useStore((state) => state.individualData);
+  const seerData = useStore((state) => state.seerData);
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState(null);
   const isFirstRender = useRef(true);
@@ -36,12 +36,15 @@ export default function IndividualDataForm({ id }) {
     setValue,
     getValues,
     watch,
+    trigger,
     formState: { errors },
   } = useForm({ defaultValues: useStore.getState().params });
 
   const inputType = watch("inputType");
   const workspaceFile = watch("workspaceFile");
   const sendNotification = watch("sendNotification");
+  const stageVariable = watch("stageVariable");
+  const selectedHeader = seerData.seerStatDictionary?.find((h) => h.name === stageVariable);
 
   const isMutatingSubmit = useIsMutating({ mutationKey: ["submitRecurrence"] });
   const isMutatingImport = useIsMutating({ mutationKey: ["importRecurrence"] });
@@ -75,7 +78,7 @@ export default function IndividualDataForm({ id }) {
   // Session rehydration
   useEffect(() => {
     if (session && !getValues("id")) {
-      setState({ individualData: session.individualData });
+      setState({ seerData: session.seerData });
       reset({ ...session.params });
     }
   }, [session]);
@@ -92,29 +95,86 @@ export default function IndividualDataForm({ id }) {
       isFirstRender.current = false;
       return;
     }
-
-    setValue("individualData", null);
+    setValue("seerStatDataFiles", null);
+    setValue("canSurvDataFile", null);
     setValue("workspaceFile", null);
-    setState({ individualData: [] });
+    setValue("stageVariable", "");
+    setValue("distantStageValue", "");
+    setState({ seerData: {} });
     setError(null);
   }, [inputType]);
 
-  async function handleIndividualDataChange(e) {
+  // Reset distantStageValue when stageVariable changes
+  useEffect(() => {
+    setValue("distantStageValue", "");
+  }, [stageVariable]);
+
+  async function handleSeerStatDataFilesChange(e) {
     const files = Array.from(e.target.files || []);
     if (!files.length) return;
+
+    const dictionaryFile = files.find((f) => /.dic$/i.test(f.name));
+    const dataFile = files.find((f) => /.txt$/i.test(f.name));
+    if (!dictionaryFile || !dataFile) {
+      trigger("seerStatDataFiles");
+      return;
+    }
 
     setIsLoading(true);
     setError(null);
     try {
-      const { data } = await parseCsvFile(files[0]);
-      setState({ individualData: data });
+      const { headers, config } = await parseSeerStatDictionary(dictionaryFile);
+      const { data } = await parseSeerStatFiles(dictionaryFile, dataFile);
+      const followUpYears = Math.min(25, getMaxFollowUpYears(config));
+
+      setState({
+        seerData: {
+          ...useStore.getState().seerData,
+          seerStatDictionary: headers,
+          seerStatData: data,
+          seerStatDataFileNames: [dictionaryFile.name, dataFile.name],
+        },
+      });
+      setValue("followUpYears", followUpYears);
+      setValue("stageVariable", "");
+      setValue("distantStageValue", "");
     } catch (err) {
       console.error(err);
-      setError("Error parsing individual data files: " + err.message);
-      setState({ individualData: [] });
+      setError("Error parsing SEER*Stat files: " + err.message);
+      setState({ seerData: {} });
     } finally {
       setIsLoading(false);
     }
+  }
+
+  async function handleCanSurvDataFileChange(e) {
+    const file = e.target.files?.[0];
+    if (!file) return;
+
+    setIsLoading(true);
+    setError(null);
+    try {
+      const { data } = await parseCsvFile(file);
+      setState({
+        seerData: {
+          ...useStore.getState().seerData,
+          canSurvData: data,
+          canSurvDataFileName: file.name,
+        },
+      });
+    } catch (err) {
+      console.error(err);
+      setError("Error parsing CanSurv file: " + err.message);
+    } finally {
+      setIsLoading(false);
+    }
+  }
+
+  function getMaxFollowUpYears(config) {
+    const sessionOptions = config["Session Options"];
+    const numberOfIntervals = +sessionOptions?.NumberOfIntervals || 30;
+    const monthsPerInterval = +sessionOptions?.MonthsPerInterval || 12;
+    return Math.ceil((monthsPerInterval * numberOfIntervals) / 12);
   }
 
   async function onSubmit(formData) {
@@ -130,8 +190,11 @@ export default function IndividualDataForm({ id }) {
     const params = {
       ...formData,
       id: newId,
+      distantStageValue: Number(formData.distantStageValue),
+      adjustmentFactorR: Number(formData.adjustmentFactorR),
+      followUpYears: Number(formData.followUpYears),
     };
-    submitForm.mutate({ params, data: useStore.getState().individualData });
+    submitForm.mutate({ params, data: seerData });
     reset(params);
     setState({ params });
     router.push(`${pathname}?id=${newId}`, { shallow: true });
@@ -148,14 +211,15 @@ export default function IndividualDataForm({ id }) {
 
   function onReset(e) {
     e.preventDefault();
-    router.push("/recurrence/individual", { shallow: false });
+    router.push("/recurrence", { shallow: false });
     reset(defaultParams);
     resetStore();
     queryClient.invalidateQueries();
     setError(null);
   }
 
-  const hasIndividualData = individualData.length > 0;
+  const hasSeerData = Object.keys(seerData).length > 0;
+  console.log(seerData);
   return (
     <div style={{ position: "relative" }}>
       <LoadingOverlay isVisible={isLoading} message="Loading data..." />
@@ -172,7 +236,7 @@ export default function IndividualDataForm({ id }) {
                 resetStore();
               }}
               disabled={!!id}>
-              <option value="data">Individual Data Files</option>
+              <option value="data">SEER*Stat/CanSurv Files</option>
               <option value="zip">Workspace (.zip)</option>
             </Form.Select>
           </Form.Group>
@@ -192,24 +256,64 @@ export default function IndividualDataForm({ id }) {
 
           {inputType === "data" && (
             <>
-              <Form.Group className="mb-4" controlId="individualDataFile">
-                <Form.Label className="required fw-bold">Individual Data File (.csv)</Form.Label>
+              <Form.Group className="mb-4" controlId="seerStatDataFiles">
+                <Form.Label className="required fw-bold">SEER*Stat Dictionary/Data Files (.dic/.txt)</Form.Label>
                 <FileInput
                   control={control}
-                  name="individualDataFile"
-                  accept=".csv"
-                  rules={{ required: "This field is required." }}
-                  onChange={handleIndividualDataChange}
-                  disabled={hasIndividualData}
+                  name="seerStatDataFiles"
+                  accept=".dic,.txt"
+                  multiple
+                  rules={{
+                    required: "This field is required.",
+                    validate: (files) => {
+                      const arr = Array.from(files || []);
+                      return (
+                        (arr.some((f) => /.dic$/i.test(f.name)) && arr.some((f) => /.txt$/i.test(f.name))) ||
+                        "SEER*Stat .dic and .txt files are required."
+                      );
+                    },
+                  }}
+                  onChange={handleSeerStatDataFilesChange}
+                  disabled={hasSeerData}
                 />
-                {errors.individualDataFile && (
+                {errors.seerStatDataFiles && (
                   <Form.Text className="text-danger">
-                    {errors.individualDataFile.message || "This field is required."}
+                    {errors.seerStatDataFiles.message || "This field is required."}
                   </Form.Text>
                 )}
               </Form.Group>
 
-              {!hasIndividualData && (
+              <Form.Group className="mb-4" controlId="canSurvDataFile">
+                <Form.Label className="required fw-bold">CanSurv Data File (.csv)</Form.Label>
+                <FileInput
+                  control={control}
+                  name="canSurvDataFile"
+                  accept=".csv"
+                  rules={{ required: "This field is required." }}
+                  onChange={handleCanSurvDataFileChange}
+                  disabled={hasSeerData}
+                />
+                {errors.canSurvDataFile && (
+                  <Form.Text className="text-danger">{errors.canSurvDataFile.message}</Form.Text>
+                )}
+              </Form.Group>
+
+              {hasSeerData && seerData.seerStatDataFileNames && (
+                <div className="mb-3">
+                  <small className="text-muted">
+                    <b>SEER*Stat:</b> {seerData.seerStatDataFileNames.join(", ")}
+                  </small>
+                  {seerData.canSurvDataFileName && (
+                    <div>
+                      <small className="text-muted">
+                        <b>CanSurv:</b> {seerData.canSurvDataFileName}
+                      </small>
+                    </div>
+                  )}
+                </div>
+              )}
+
+              {!hasSeerData && (
                 <div className="mb-3">
                   <Button
                     className="p-0"
@@ -223,13 +327,22 @@ export default function IndividualDataForm({ id }) {
                           const blob = await response.blob();
                           return new File([blob], fileName, { type: blob.type });
                         };
-
-                        const individualDataFile = await fetchFile(
-                          "/data/recurrence_risk_individualdata.csv",
-                          "recurrence_risk_individualdata.csv"
+                        const dicFile = await fetchFile(
+                          "/data/recurrence_risk_groupdata/groupdata_example_seer.dic",
+                          "groupdata_example_seer.dic"
                         );
-                        setValue("individualDataFile", asFileList([individualDataFile]));
-                        await handleIndividualDataChange({ target: { files: [individualDataFile] } });
+                        const txtFile = await fetchFile(
+                          "/data/recurrence_risk_groupdata/groupdata_example_seer.txt",
+                          "groupdata_example_seer.txt"
+                        );
+                        const cansurvFile = await fetchFile(
+                          "/data/recurrence_risk_groupdata/groupdata_example_cansurv.csv",
+                          "groupdata_example_cansurv.csv"
+                        );
+                        setValue("seerStatDataFiles", asFileList([dicFile, txtFile]));
+                        setValue("canSurvDataFile", asFileList([cansurvFile]));
+                        await handleSeerStatDataFilesChange({ target: { files: [dicFile, txtFile] } });
+                        await handleCanSurvDataFileChange({ target: { files: [cansurvFile] } });
                       } catch (err) {
                         console.error("Error loading example files:", err);
                       }
@@ -241,9 +354,9 @@ export default function IndividualDataForm({ id }) {
                       className="p-0"
                       variant="link"
                       size="sm"
-                      href="/data/recurrence_risk_individualdata.csv"
+                      href="/data/recurrence_risk_groupdata/recurrence_risk_groupdata.zip"
                       download>
-                      Download Example (.csv)
+                      Download Example (.zip)
                     </Button>
                   </div>
                 </div>
@@ -252,13 +365,84 @@ export default function IndividualDataForm({ id }) {
           )}
         </fieldset>
 
-        {hasIndividualData && (
+        {hasSeerData && (
           <fieldset className="fieldset shadow-sm border rounded my-4 pt-4 px-3">
             <legend className="legend fw-bold">Parameters</legend>
+
+            <Form.Group className="mb-4" controlId="stageVariable">
+              <Form.Label className="required fw-bold">Stage Variable</Form.Label>
+              <Form.Select
+                {...register("stageVariable", { required: "This field is required." })}
+                isInvalid={!!errors.stageVariable}>
+                <option value="" hidden>
+                  No stage variable selected
+                </option>
+                {seerData.seerStatDictionary
+                  ?.filter((h) => h.factors?.length > 0)
+                  .map((h) => (
+                    <option key={h.name} value={h.name}>
+                      {h.name}
+                    </option>
+                  ))}
+              </Form.Select>
+              <Form.Control.Feedback type="invalid">{errors.stageVariable?.message}</Form.Control.Feedback>
+            </Form.Group>
+
+            <Form.Group className="mb-4" controlId="distantStageValue">
+              <Form.Label className="required fw-bold">Distant Stage Value</Form.Label>
+              <Form.Select
+                {...register("distantStageValue", { required: "This field is required." })}
+                isInvalid={!!errors.distantStageValue}>
+                {!stageVariable && (
+                  <option value="" hidden>
+                    No distant stage value selected
+                  </option>
+                )}
+                {selectedHeader?.factors?.map((factor) => (
+                  <option key={factor.value} value={factor.value}>
+                    {factor.value} - {factor.label}
+                  </option>
+                ))}
+              </Form.Select>
+              <Form.Control.Feedback type="invalid">{errors.distantStageValue?.message}</Form.Control.Feedback>
+            </Form.Group>
+
+            <Form.Group className="mb-4" controlId="adjustmentFactorR">
+              <Form.Label className="required fw-bold">Adjustment Factor r</Form.Label>
+              <Form.Control
+                {...register("adjustmentFactorR", {
+                  required: "This field is required.",
+                  valueAsNumber: true,
+                  min: { value: 0.5, message: "Please enter a value equal to or greater than 0.5." },
+                  max: { value: 2, message: "Please enter a value equal to or less than 2." },
+                })}
+                type="number"
+                step="0.1"
+                min="0.5"
+                max="2"
+                isInvalid={!!errors.adjustmentFactorR}
+              />
+              <Form.Control.Feedback type="invalid">{errors.adjustmentFactorR?.message}</Form.Control.Feedback>
+            </Form.Group>
+
+            <Form.Group className="mb-4" controlId="followUpYears">
+              <Form.Label className="required fw-bold">Years of Follow-up</Form.Label>
+              <Form.Control
+                {...register("followUpYears", {
+                  required: "This field is required.",
+                  valueAsNumber: true,
+                  min: { value: 1, message: "Please enter a value equal to or greater than 1." },
+                })}
+                type="number"
+                min="1"
+                isInvalid={!!errors.followUpYears}
+              />
+              <Form.Control.Feedback type="invalid">{errors.followUpYears?.message}</Form.Control.Feedback>
+            </Form.Group>
           </fieldset>
         )}
 
-        {hasIndividualData && (
+        {hasSeerData && (
           <fieldset className="fieldset shadow-sm border rounded my-4 pt-4 px-3">
             <legend className="legend fw-bold">Notifications</legend>
             <Form.Group className="mb-3">
@@ -306,7 +490,7 @@ export default function IndividualDataForm({ id }) {
             disabled={
               inputType === "zip"
                 ? !Array.from(workspaceFile || []).length || !!isSubmitting
-                : !hasIndividualData || !!isSubmitting
+                : !hasSeerData || !!isSubmitting
             }>
             {isSubmitting ? (
               <>
